requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.25.2
Django==5.1.2
django-cors-headers==4.5.0
django-rest-framework-nested==0.0.1
djangorestframework==3.15.2
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.1
